Fixes to bytes/unicode strings encoding and decoding, errors="ignore", see the UNICODE_ENCODE_ERRORS and BYTES_DECODE_ERRORS global variables in string_utils.pyx.
Fixed Frame.ExecuteJavascript(), if the string contained
utf-8 characters it would probably result in errors.

Updated ExceptHook in the examples to handle UnicodeDecodeError
exceptions when writing to the log file or printing invalid
characters.

Added PY_MAJOR_VERSION compile time constant.

Changed the c string types in Cython to be always 'bytes'
with the 'utf-8' encoding, any decoding to unicode should
be explicit.

Updated the way the cefpython module is imported in the examples,
the previous method was treacherous.

Changed the name of the ApplicationSettings["unicode_to_bytes_encoding"]
option to "string_encoding".

Fixed importing urlparse in Python 3.

Updated CEF 1 to Chrome version 27.0.1453.110.