Improvements to Linux and Windows installers. Issue 145, Issue 105, Issue 99.
setup.py will use either distutils or setuptools depending on which
one is available. setuptools is not distributed by default.

Post install instructions will now create a symbolic link to
libudev.so.0 on the fly.

Loading libffmpegsumo.so in the browser process may cause issues.
It is designated only for the renderer subprocess.