Provide default implementation for js and file dialogs on Linux (#241)...
In wxpython.py and gtk2.py examples the js alert dialog when run
from a popup gives focus to the main window. The popup window is
created internally by CEF and this probably needs to be changed
so that popups are created by wx in OnBeforePopup and this will
resolve issue with alert focus. In qt4.py and tkinter.py and
hello_world.py examples focus works fine in the popup window.

Add --hello-world flag to build.py and run_examples.py tools.