Allow optionally binding all properties and methods of an object
When binding an object with SetObject(name, obj, allow_properties=True) all methods and properties that
are allowed with IsValueAllowedRecursively() and don't start with "__" will be included in the binding.
Note that properties will be bound by copy, there will be no automatic updating of js <-> py objects when
one is modified.