Further changes to message loop issues on Mac (#442).
Minor fix to CefAppProtocol implementation, see CEF Forum
topic for details.

Updated Qt example so that message loop work is not
called in a timer when message pump is enabled. This
is not required.

In wxPython apps it is still required to enable both
message pump and message looper timer work.